:recycle:  moving more components out of common.tsx
